+++ example/main.dart
@@ -1,8 +1,7 @@
-import 'package:http/http.dart';
 import 'package:mpesa_sdk_dart/mpesa_sdk_dart.dart';
 
 main() async {
-  String? token = MpesaConfig.getBearerToken(
+  String token = MpesaConfig.getBearerToken(
     'API_KEY_HERE',
     'PUBLIC_KEY_HERE',
   );
@@ -12,11 +11,10 @@ main() async {
   PaymentRequest payload = PaymentRequest(
     inputTransactionReference: 'T12344C',
     inputCustomerMsisdn: '25884xxxxxxx',
-    inputAmount: 10,
+    inputAmount: 10.0,
     inputThirdPartyReference: '11114',
     inputServiceProviderCode: '171717', // business short code
   );
 
-  Response response = await MpesaTransaction.c2b(token, apiHost, payload);
-  print(response.body);
+  await MpesaTransaction.c2b(token, apiHost, payload);
 }